Pull request #3689: [MCSE-1156] [NVM] Fix MISRA rule 10.4 violation
Merge in CONNINT/bluetooth_public from mcse-1156_misra_required to develop/main

Squashed commit of the following:

commit 2d4dbd5e50cc03f8b701ffad516cf394dbfac235
Author: Philippe Guasch <philippe.guasch@nxp.com>
Date:   Mon Oct 27 17:21:13 2025 +0100

    [MCSE-1156] [NVM] Fix MISRA rule 10.4 violation
    
    - gNvRecordsCopiedBufferSize_c must be specified as unsigned

+++ boards/frdmmcxw71/app_preinclude_common.h
@@ -217,7 +217,7 @@
     #define  gUnmirroredFeatureSet_d             (1)
     #if gNvFragmentation_Enabled_d
         /* Buffer size large enough to accommodate the maximum number of CCCDs for every device. */
-        #define  gNvRecordsCopiedBufferSize_c    (gMaxBondedDevices_c * 16)
+        #define  gNvRecordsCopiedBufferSize_c    ((uint16_t)gMaxBondedDevices_c * 16u)
     #endif
 #endif
